Correcting .txt file table output in the reanalysis folder
Upon notification that the three merge_fix.txt tables in the reanalysis_output folder were the same as those in the replication folder but with different titles, I re-ran the reanalysis.R code stored in this repo.

While the correct tables had been sent in the original memo to the editors and authors, it is worth correcting these plain text versions in the public repository, and ensuring their filenames reflect the current version of code. This commit also removes the incorrect txt tables and edits the readme on the homepage.